refractoring server to module.exports
